Feat: RecommendedSpaceCard 뷰 모델 분리
뷰 모델을 분리하여 관리를 용이하게 함
